Slightly change in application multiple runs macros
The ideia is that instead of calling the kernel function IN_RUN times,
re-using allocated buffers and their respective initialization, the
pre-kernel setup is also executed before each kernel run, as an attempt
to ensure cold cache. The new macros allow such scenario, but at the
time of this commit, the kernels are configured to run only once. This
is only needed to vtune profilling in fast kernels, so there's not much
need for tweaking the existing codebase