setup: Add custom-GNU compilation support
For OpenMP offloading, GNU compilers require specific '-foffload'
options that are not suported in e.g. Clang. Therefore, compiler
specific flags are stored in compiler specific variables:
GCC_OFFLOAD_FLAGS. It should be included in the compiling command only
and only if the compiler is gcc/g++. Therefore, a new flag is introduced
to tell the current compiler being used