fix: Add workaround to prevent dead-code elimination
I noticed that Clang is quite aggressive with some optimizations. It
turns out it was eliminating kernel's loops because the computation
results were not used afterwards. This commit uses the same technique as
in original Polybench source-code, where it iterates over the arrays and
prints values. However, to not spam the stdout, it actually accumulates
the values and prints the final sum. The value is also sent to stderr to
avoid problems with existing scripts